+++ crates/javy/src/lib.rs
@@ -40,7 +40,7 @@
 //! * `messagepack` - functions for converting between [`quickjs::JSValueRef`]
 //!   and MessagePack byte slices
 
-pub use config::Config;
+pub use config::*;
 pub use rquickjs as quickjs;
 pub use runtime::Runtime;
 use std::str;
@@ -61,6 +61,8 @@ pub mod messagepack;
 
 #[cfg(feature = "json")]
 pub mod json;
+
+mod apis;
 
 /// Print the given JS value.
 ///
@@ -118,7 +120,7 @@ impl<'js> Args<'js> {
     }
 }
 
-/// Alias for `Args::hold(cx, args).release()`
+/// Alias for [`Args::hold(cx, args).release()`]
 #[macro_export]
 macro_rules! hold_and_release {
     ($cx:expr, $args:expr) => {
@@ -126,7 +128,7 @@ macro_rules! hold_and_release {
     };
 }
 
-/// Alias for [Args::hold]
+/// Alias for [`Args::hold`]
 #[macro_export]
 macro_rules! hold {
     ($cx:expr, $args:expr) => {